feat: map-sdk 경계 고정 및 테스트 추가
- Map/Context를 libs로 이동해 SDK 접근 경계를 정리
- 도메인은 re-export만 유지해 사용 방식 변화 최소화
- useMap, Map 테스트 작성
