+++ frontend/src/libs/map-sdk/index.ts
@@ -7,6 +7,10 @@ export { kakaoMapAdapter } from './adapters/kakaoMapAdapter';
 // Controllers
 export { mapController } from './controllers/mapController';
 
+// Components
+export { default as Map } from './components/Map/Map';
+export type { MapProps, LatLng } from './components/Map/Map.types';
+
 // Hooks
 export { useMapSdkLoader } from './hooks/useMapSdkLoader';
 export type { UseMapSdkLoaderReturn } from './hooks/useMapSdkLoader';
